Summary of changes from v2.4.19-rc1 to v2.4.19-rc2 ============================================
<davem@redhat.com> (02/06/24 1.553)
	[PATCH] Fix tg3 oops after 2 billion packets

<trini@kernel.crashing.org> (02/06/24 1.554)
	[PATCH] Fix building with 'make -r -R'

<stelian.pop@fr.alcove.com> (02/06/24 1.555)
	[PATCH] 2.4.19-pre10 do_mounts.c compile fix

<mikulas@artax.karlin.mff.cuni.cz> (02/06/25 1.556)
	[PATCH] Fix HPFS NULL pointer access

<kaos@ocs.com.au> (02/06/25 1.557)
	[PATCH] 2.4.19-rc1 drivers/scsi/ips.c needs __devexit_p

<kaos@ocs.com.au> (02/06/25 1.558)
	[PATCH] 2.4.19-rc1 Remove warning on make modules_install

<eike@bilbo.math.uni-mannheim.de> (02/06/25 1.559)
	[PATCH] typo in drivers/scsi/sim710.c

<sunil.saxena@intel.com> (02/06/25 1.560)
	[PATCH] P4 illegal instruction failures fixes

<dwmw2@redhat.com> (02/06/25 1.561)
	[PATCH] JFFS2 rename()-over-victim fix.

<bunk@fs.tum.de> (02/06/25 1.562)
	[PATCH] fix .text.exit error in cpqarray

<bunk@fs.tum.de> (02/06/25 1.563)
	[PATCH] fix .text.exit errors in orinoco_{pci,plx}.c

<bunk@fs.tum.de> (02/06/25 1.564)
	[PATCH] fix .text.exit error in neofb.c

<bunk@fs.tum.de> (02/06/25 1.565)
	[PATCH] fix .text.exit error in mtd/maps/pci.c

<trond.myklebust@fys.uio.no> (02/06/25 1.566)
	[PATCH] 2.4.19-rc1 Fix NFS attribute caching bug

<benh@kernel.crashing.org> (02/06/26 1.567)
	[PATCH] PPC 603 CPU fix 1/4

<benh@kernel.crashing.org> (02/06/26 1.568)
	[PATCH] PPC 603 CPU fix 2/4

<benh@kernel.crashing.org> (02/06/26 1.569)
	[PATCH] PPC 603 CPU fix 3/4

<benh@kernel.crashing.org> (02/06/26 1.570)
	[PATCH] PPC i2c-keywest typo fix

<garloff@suse.de> (02/06/26 1.571)
	[PATCH] Additional BLIST_LARGELUN devices

<maxk@qualcomm.com> (02/06/26 1.572)
	[PATCH] Important Bluetooth fixes

<schwidefsky@de.ibm.com> (02/06/26 1.573)
	[PATCH] 2.4.19-rc1: important s390 fixes.

<ladis@psi.cz> (02/06/27 1.574)
	[PATCH] radio-sf16fmi: fix typo

<neilb@cse.unsw.edu.au> (02/06/27 1.575)
	[PATCH] PATCH - fix for raid1 superblock update in 2.4.19-rc1

<hugh@veritas.com> (02/06/27 1.576)
	[PATCH] Fix shm_destroy lock hang

<garloff@suse.de> (02/06/27 1.577)
	[PATCH] Remove mistaken BLIST_LARGELUN flag from Chaparral device

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.1)
	MIPS update; sync with 2.4.19-rc1.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.2)
	Resync MIPS port with 2.4.19-rc1, updates from oss cvs.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.3)
	HP Laserjet port updates.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.4)
	Updates for MIPS, Inc. evaluation boards.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.5)
	SGI IP22 updates; sync with Linux 2.4.19-rc1.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.6)
	DEC Updates

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.7)
	The FPU emulator is no longer a compile option so remove documentation

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.8)
	SGI IP27 updates.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.9)
	Collection of random MIPS networking updates.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.10)
	Random updates for Cobalt MIPS boxen.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.11)
	Update of SB1250 / BCM1250 SOC support.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.12)
	Update support for Galileo MIPS evaluation boards.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.13)
	Devfs fixes for IP22 serial driver.

<ralf@dea.linux-mips.net> (02/06/28 1.555.1.14)
	Sound updates for various MIPS platforms.

<elenstev@mesatop.com> (02/06/28 1.578)
	[PATCH] compile fix for 2.5 kdev_t compatibility macros

<hch@infradead.org> (02/06/28 1.579)
	[PATCH] backout disk statistics in /proc/partitions

<Nicolas.Aspert@epfl.ch> (02/06/28 1.580)
	[PATCH] Make agpgart module able to determine aperture size of i815

<hch@infradead.org> (02/06/28 1.581)
	[PATCH] fix up 802 template code

<marcelo@plucky.distro.conectiva> (02/06/28 1.582)
	Backout readv/writev SuS compliance: it was breaking oSF bins on Alpha

<p2@ace.ulyssis.sutdent.kuleuven.ac.be> (02/07/02 1.583)
	[PATCH] fix typo in tulip net driver

<rmk@arm.linux.org.uk> (02/07/02 1.584)
	[PATCH] fix orinoco wireless system-freeze problem

<jakob.kemi@telia.com> (02/07/02 1.585)
	[PATCH] w9966 egcs compile fix

<jgarzik@mandrakesoft.com> (02/07/03 1.586)
	Remove starfire_firmware.pl by request, as it enables

<alan@lxorguk.ukuu.org.uk> (02/07/05 1.585.1.1)
	[PATCH] Re add readv/writev SuS compliance

<ink@jurassic.park.msu.ru> (02/07/05 1.585.1.2)
	[PATCH] Implement Alpha  osf_readv/osf_writev

<twaugh@redhat.com> (02/07/05 1.588)
	[PATCH] 2.4.19-rc1: ppdev read(...,0) fix

<bunk@fs.tum.de> (02/07/05 1.589)
	[PATCH] fix .text.exit error in ieee1394/ohci1394.c

<trond.myklebust@fys.uio.no> (02/07/10 1.590)
	[PATCH] Fix NFS lockd oops

<rusty@rustcorp.com.au> (02/07/10 1.591)
	[PATCH] Fix SMP compilation with 'multiquad' driver

<rusty@rustcorp.com.au> (02/07/10 1.592)
	[PATCH] utimes permission check

<rusty@rustcorp.com.au> (02/07/10 1.593)
	[PATCH] PATCH 2.5: update CREDITS

<rusty@rustcorp.com.au> (02/07/10 1.594)
	[PATCH] Re: Trivial fix for _proc operation: Thankyou for your

<alan@lxorguk.ukuu.org.uk> (02/07/11 1.595)
	[PATCH] Fix several pdc202xx problems

<marcelo@plucky.distro.conectiva> (02/07/12 1.596)
	Fix race between writepage() and fsync() -- Kudos goes to Andrea

<swanson@uklinux.net> (02/07/12 1.597)
	[PATCH] Fix radeon DRI lockups

<elenstev@mesatop.com> (02/07/12 1.598)
	[PATCH] 2.4.19-rc1, 38 Configure.help texts

<viro@math.psu.edu> (02/07/13 1.613)
	[PATCH] readdir

<rusty@rustcorp.com.au> (02/07/15 1.614)
	[PATCH] Overrun for SMP connection tracking under load

<marcelo@plucky.distro.conectiva> (02/07/16 1.615)
	Fix up mispatching of trivial proc_dointvec_max fix.

<mostrows@speakeasy.net> (02/07/16 1.616)
	[PATCH] oops fix in PPPoE

<marcelo@plucky.distro.conectiva> (02/07/16 1.617)
	Additional BLIST_LARGELUN device

<marcelo@plucky.distro.conectiva> (02/07/16 1.618)
	Changed EXTRAVERSION to -rc2

<bcollins@debian.org> (02/07/16 1.619)
	[PATCH] Very important IEEE1394 fixed

<bunk@fs.tum.de> (02/07/16 1.620)
	Add []checkout:get to BK config file

<marcelo@plucky.distro.conectiva> (02/07/16 1.621)
	Remove IO stats headers /proc/statistics

+++ include/net/bluetooth/hci_core.h
@@ -23,7 +23,7 @@
 */
 
 /* 
- * $Id: hci_core.h,v 1.3 2002/04/17 18:55:21 maxk Exp $ 
+ * $Id: hci_core.h,v 1.4 2002/06/25 22:04:43 maxk Exp $ 
  */
 
 #ifndef __HCI_CORE_H
@@ -339,8 +339,8 @@ static inline void hci_sched_tx(struct hci_dev *hdev)
 /* ----- HCI protocols ----- */
 struct hci_proto {
 	char 		*name;
-	__u32		id;
-	__u32		flags;
+	unsigned int 	id;
+	unsigned long	flags;
 
 	void		*priv;
